Add sample loader service with 100-row limit
